tests file OK

+++ Eval_Opti_Solution/Multi-Objectifs/Arc.h
@@ -8,5 +8,5 @@ public:
 	double longueur;
 	double cout;
 
-	Arc(int num, int f, int t, int l, int c) : numero_arc(num), from(f), to(t), longueur(l), cout(c) {}
+	Arc(int num, int f, int t, double l, double c) : numero_arc(num), from(f), to(t), longueur(l), cout(c) {}
 };
